requirements.txt
feedparser==6.0.11
requests==2.32.3
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0
duckduckgo-search>=4.0.0
python-dotenv>=1.0.0
